game/map: Added pawn enemy

+++ game/inc/tilemap/tilemap.h
@@ -34,7 +34,7 @@ inline void TILEMAP_update(const TileMap *const tilemap) {
       mapLevelY, 0, 0, tilemap->w, tilemap->h, DMA);
 }
 
-inline void TILEMAP_updateRightTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
+inline void TILEMAP_updateRightTile(s16 x, s16 y, s16 mapPosX, s16 mapPosY,
                                     u16 tileIndex) {
   VDP_setTileMapXY(TILEMAP_PLANE,
                    TILE_ATTR_FULL(TILEMAP_PAL, 0, FALSE, FALSE, tileIndex),
@@ -50,7 +50,7 @@ inline void TILEMAP_updateRightTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
                    x + mapPosX + 3, y + mapPosY + 1);
 }
 
-inline void TILEMAP_updateLeftTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
+inline void TILEMAP_updateLeftTile(s16 x, s16 y, s16 mapPosX, s16 mapPosY,
                                    u16 tileIndex) {
   VDP_setTileMapXY(TILEMAP_PLANE,
                    TILE_ATTR_FULL(TILEMAP_PAL, 0, FALSE, FALSE, tileIndex),
@@ -66,7 +66,7 @@ inline void TILEMAP_updateLeftTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
                    x + mapPosX - 2, y + mapPosY + 1);
 }
 
-inline void TILEMAP_updateUpTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
+inline void TILEMAP_updateUpTile(s16 x, s16 y, s16 mapPosX, s16 mapPosY,
                                  u16 tileIndex) {
   VDP_setTileMapXY(TILEMAP_PLANE,
                    TILE_ATTR_FULL(TILEMAP_PAL, 0, FALSE, FALSE, tileIndex),
@@ -82,7 +82,7 @@ inline void TILEMAP_updateUpTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
                    x + mapPosX, y + mapPosY - 2);
 }
 
-inline void TILEMAP_updateBottomTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
+inline void TILEMAP_updateBottomTile(s16 x, s16 y, s16 mapPosX, s16 mapPosY,
                                      u16 tileIndex) {
   VDP_setTileMapXY(TILEMAP_PLANE,
                    TILE_ATTR_FULL(TILEMAP_PAL, 0, FALSE, FALSE, tileIndex),
@@ -98,7 +98,7 @@ inline void TILEMAP_updateBottomTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
                    x + mapPosX, y + mapPosY + 3);
 }
 
-inline void TILEMAP_updateUpRighTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
+inline void TILEMAP_updateUpRighTile(s16 x, s16 y, s16 mapPosX, s16 mapPosY,
                                      u16 tileIndex) {
   VDP_setTileMapXY(TILEMAP_PLANE,
                    TILE_ATTR_FULL(TILEMAP_PAL, 0, FALSE, FALSE, tileIndex),
@@ -114,7 +114,7 @@ inline void TILEMAP_updateUpRighTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
                    x + mapPosX + 3, y + mapPosY - 1);
 }
 
-inline void TILEMAP_updateUpLeftTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
+inline void TILEMAP_updateUpLeftTile(s16 x, s16 y, s16 mapPosX, s16 mapPosY,
                                      u16 tileIndex) {
   VDP_setTileMapXY(TILEMAP_PLANE,
                    TILE_ATTR_FULL(TILEMAP_PAL, 0, FALSE, FALSE, tileIndex),
@@ -130,7 +130,7 @@ inline void TILEMAP_updateUpLeftTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
                    x + mapPosX - 1, y + mapPosY - 1);
 }
 
-inline void TILEMAP_updateBottomRightTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
+inline void TILEMAP_updateBottomRightTile(s16 x, s16 y, s16 mapPosX, s16 mapPosY,
                                           u16 tileIndex) {
   VDP_setTileMapXY(TILEMAP_PLANE,
                    TILE_ATTR_FULL(TILEMAP_PAL, 0, FALSE, FALSE, tileIndex),
@@ -146,7 +146,7 @@ inline void TILEMAP_updateBottomRightTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
                    x + mapPosX + 3, y + mapPosY + 3);
 }
 
-inline void TILEMAP_updateBottomLeftTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
+inline void TILEMAP_updateBottomLeftTile(s16 x, s16 y, s16 mapPosX, s16 mapPosY,
                                          u16 tileIndex) {
   VDP_setTileMapXY(TILEMAP_PLANE,
                    TILE_ATTR_FULL(TILEMAP_PAL, 0, FALSE, FALSE, GREEN_TILE),
@@ -162,7 +162,8 @@ inline void TILEMAP_updateBottomLeftTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY,
                    x + mapPosX - 2, y + mapPosY + 3);
 }
 
-inline void TILEMAP_setTile(s8 x, s8 y, u8 mapPosX, u8 mapPosY, u16 tileIndex) {
+inline void TILEMAP_setTile(s16 x, s16 y, s16 mapPosX, s16 mapPosY,
+                            u16 tileIndex) {
   VDP_setTileMapXY(TILEMAP_PLANE,
                    TILE_ATTR_FULL(TILEMAP_PAL, 0, FALSE, FALSE, tileIndex),
                    x + mapLevelX, y + mapLevelY);
